Logging and a few minor edits
I made every control I could find that does anything create a one-line addition to the log file. To do this, I added onclick, onchange or onblur events to all of the affected controls, and added ids to a few. Most of the changes are in app.js, where I placed the methods which create the log entries. I also corrected a few typos and added a few comments.