Simplified the ID system
I set idrs for all the elements we need to access within each widget, and removed their sequential absolute IDs. That means that we can eliminate the idReplace method, and that widget ids are now sequential. I wrote a new method, app.getChildByIdr(element, idr) which finds the first descendant of the element which has the given idr.