Added recording and replay
I added buttons to record/stop recording, download a recorded file, upload a recorded file, play a recorded file, and clear the DB and web page. I placed the variables and methods required to run these buttons in a new class, regressionTesting.js, and created an instance of that class in the app object when the page loads.