Several minor changes and cleaned up code
I moved all the javascript to a new class, dragDrop, and created an instance of it in 1-drag.html. I added a few comments to make the code clearer for me and hopefully for anyone else who reads it. I changed the functionality in the following ways:

The word "Insert" can no longer be edited.

The insert text box is autofocused when the page loads and when you finish editing an existing node

When you open a node to edit, the text it contains is selected

Changed the save() method for editing from onchange to onblur, so if you change your mind you can stop editing without making a change

Hitting enter while editing now saves, just like losing focus does