More tinkering
Added a show/hide button for the Insert text box. I managed to set it up so that hiding the text box with data in it does not add the data to the list, and when you show the text box again, it still has the data in it and gets focus again. I'm not really happy with HOW I did that, though, so I wouldn't mind going over it again. I also changed the recordToggle function in regressionTesting so that the name of the file it creates is a little more informative.