Added recording to drag and other changes
I rewrote 1-Drag.html and dragDrop.js so that the onclick, ondrag, etc. events for the show/hide button, the insert field and all the draggable fields were added by JS. I added the "widget" class to the container object so widgetGetId would work on it, and added idrs to everything inside the container. I implemented recording and playback on 1-drag.html and made three recordings with descriptive names. I moved both drag-drop files and all three recordings to 4-regressionTests. I rewrote the code to simulate an action in regressionTesting.js because I found an easier and more versatile way to do it.