Bug Fixes
These are three bugs I found yesterday and fixed. Luckily I wrote down what I did, so when I started over today I decided to re-fix them. It looks like several lines got erased from widgetNode.buildWidget, which would have closed the td, tr and tbody, as well as creating the "to" element needed for toDOM to create the right-hand relationship table. I replaced the missing code. Result: Opening a node for editing now has a relationship table on either side, rather than just one on the left, and does not produce an error message.

I also changed two references to this.data, which doesn't exist in widgetNode.js, to this.dataNode, which does. This allows data to show up in a node which is opened for editing.

Finally, in widgetTableQuery, I rewrote the code to open a trash node for editing. It seemed to have been copied from widgetTableNodes, and referred to this.queryObject and to a result's "n" value, neither of which existed in widgetTableQuery. I got the label from element.nextElementSibling instead (because the label comes right after the ID in the Trash table), and added n, the whole node, to the list of things for the trash query to return.

I also deleted the two classes that you said we wouldn't need anymore, widgetGraph and widgetRelation, and the extra initialization of this.data in db.js