Tinkering with regressionTesting
Added more functionality to regressionTesting class (play back steps one at a time, delay between steps). Added function to regressionTesting which creates testing inputs automatically if given the name of the global variable to use in the onclick functions, and an element named "regressionHeader" to place them in. Updated all the HTML files in 4-regressionTests to pass in the global variable's name and have a regressionHeader div in order to take advantage of the new function. Also, changed my neo4j password - will need to coordinate to get our passwords the same again.

The changes to regressionTesting didn't fix the timing bug with editing nodes, but they did make it easy to work around it. I'll work more on the test files tomorrow.