Starting to work on relation table
Introduced “editable” tag for template in dragDrop and dragDropTable. Only elements with “editable” tag have an input added to them. For the time being, all attributes of the template, other than id, are also duplicated in the input and finished rows.

Rewrote dragDrop and tableDragDrop to take the global variable as its last argument for the time being, so I can use them in _app.html. If no third argument is provided, it assumes dragDrop itself is the global variable, so the two dragDrop HTML files don’t need changing. This is obviously kludgy, but it doesn’t need to be permanent - it’ll get fixed when we switch to using app.widget more universally, but that will require extensive rewrites in an area we’re not prioritizing right now.

Rewrote widgetRelation and widgetNode so that only ONE dragDrop table is created, with a single editable column (Comment), and that column can be added to, edited and reordered. The other columns are blank for now. It should work fine with the existing rows that are added for existing relations - you should still be able to drag, drop and edit those.

Deleted the two redundant copies of dragDropTable (tableDragDrop and dragDroptTable) from 4-regressionTests/LV4. Changed “var” to “let” and, when possible, “let” to “const” in whatever code I was working on.