Made widgets more consistent; removed references to global vars other than app
Rewrote code for dragDrop, widgetRelations and regressionTesting classes so that instances of those classes are treated like any other widget - the JS object is stored in app.widgets, the class of the element holding everything is "widget" and the id of that element matches the key used in app.widgets, and everything within that widget element that the user could interact with has an idr. That means that we can now use app.widget() to trigger their functions, so we don't need to refer so much to global variables. We can also use app.widgetGetId() on elements in those widgets now. And since the elements within the widget will use idrs, not ids, we don't need to worry about keeping them unique outside the context of that individual widget.

I also removed multiple regressionTesting objects because logging won't work properly unless ALL steps are recorded by the same object. Now app creates its own copy of regressionTesting in its constructor and everything else uses app's.

I expanded app.widget() slightly so that you can now use it to call functions that take an argument other than the element which triggered the call. For instance, several of dragDrop's functions need the element AND the event which triggered the call, and app.widget can now pass both on to the desired function. I think I've finally finished the rewrites we talked about a few days ago and can go back to working on getting the dragDrop table in widgetRelations to work.

One downside to the work I've done is that since some elements' IDs and/or IDRs have changed, the regression test recordings I've already made are unlikely to run and even less likely to report success. I'll rerecord them as soon as I get a chance.