Working on widgetRelation dragDrop table
I think I got the dragDrop table working. I already had it where you could add a new comment and edit a comment. Now you can drag a node ID from a widgetTableNodes table into the N# column, and that ID will be added to that column. You can click the "Delete" button (I changed it from "X" to make it a little clearer), and it will delete a new (unsaved) node or mark an existing one for deletion. When a node is marked for deletion, the "Delete" button becomes a "Restore" button. Nodes are marked with classes which determine both their color and the way they are processed when the "Save and Sync" button is clicked. Deleted lines are red, changed ones are yellow and new ones are green. Clicking "Save and Sync" adds and removes relations as necessary (it's not possible to change the start and end points of a relation, so for the time being I'm mimicking it by deleting the existing relation and creating a new one). I also edited db so that it's possible to run a query without passing in an object and object method.