Cleaned up code; fixed toggle button; added "end" relations table; made relation tables appear when a new node is saved
I removed the app functions widgetNodeNew, widgetRelationNew and dragDropTableNew, instead adding the new widgets to app.widgets when they're created by widgetTableNodes, widgetTableQuery, widgetNode and widgetRelation.

I expanded the "orderby" string for each node in metadata so that the ordering for all the displayed fields is defined.

I realized that the toggle button in widgetRelation was only showing/hiding the "Save and Sync" button, so I checked the code and found that it was only affecting its next sibling. I rewrote the code so that it affects all its siblings that come after it.

I added a switch to widgetRelations.complete so that instead of always displaying the end node of a relation (which only works for the "start" table), it displays the start or end node depending on the relationship type.

I moved the code in widgetNode which creates a new widgetRelations to its own function and added a second line to it to create the "end" table. I called that function from the constructor if data was provided, and also from addComplete so that once a new node is added, it becomes possible to add relations to it.

I added a few comments about code that doesn't seem to be used, and deleted a few lines of comments that were out of place - it looked like notes on how dragDrop inputs work, but they were in widgetRelations.complete, which doesn't even create inputs.