Mostly working on improving dragDrop table; moved some code around
First, the small changes. I realized I should have had the code modifying app.widgets in constructors rather than the code that calls them, so I moved it. I moved the definition of dragDrop.dropData to app.relations and added it to the dragDrop tables after widgetRelations created them, since its effects are pretty specific to what widgetRelations needs the table to do. Similarly, I removed the code which always adds the changedData class from dragDrop.save, and instead wrote a function to do that called changeComment in widgetRelations and added it to the newly-created dragDrop tables. I realized it wasn't necessary to pass the widgetID to dragDrop or dragDropTable if it was going to have to interact directly with app.idCounter anyway, so I removed it from the arguments to dragDrop and dragDropTable. I added "name" to the list of attributes of an address in metadata because we discussed making every node type have a name and displaying the name in the widgetRelations dragDrop table. None of these changes affects the UI except for the last one, which added a "name" column to the address table and a "name" field to address nodes.

All remaining changes were made to widgetRelations, dragDrop and widgetTableNodes in order to make the changes we discussed to how widgetRelations works.

I added a "direction" key to placeholder nodes. When a node is added, if the "node ID" field in the table is blank (meaning no node has been chosen), a circular relationship is created and it's given a "direction" value that matches the widgetRelation's relationType. When widgetRelations searches for relations to or from a node, it now checks for a "direction" key. If there's a "direction" key, and it doesn't match the widget's relationType, then that relationship isn't shown at all (so a relation that's a placeholder for the "start" table doesn't show up in the "end" table. If the direction key DOES match the widget's relationType, then the relation is shown, but the nodeID is NOT shown (because it's just a placeholder- you don't ACTUALLY want a circular relationship).

I had the search for relations return the node each relation leads to as well as the relation itself, so it would be easier to access information about those nodes like their name and type. I added columns to the dragDrop tables for the node name and type.

I made node names in widgetTableNodes draggable, and gave an ondrop function to the node name cells in the dragDrop table, so that you can now drag EITHER the node ID OR the name from the widgetTableNodes table into the dragDrop table, and drop it in either the ID or the name column. I edited widgetTableNodes.drag to store a stringified object representing the node's ID, name and type, and edited the dropData function (defined in widgetRelations but used by dragDrop) to parse that data and store it all in the appropriate tables.

 I created a variable called this.existingRelations in widgetRelations. Every time you search for relations and build a dragDrop table, existingRelations is populated with data about those relations (at the moment, the relation ID, node ID, name and type of the node, and comment) and passed into the dragDrop table that is created. I modified the dropData function, which widgetRelations adds to dragDrop, to check whether the ID, name and type are different from recorded values and only mark them as changed if they are. I also added logic to the changeComment function to only mark an edited comment as "changed" if it doesn't match the recorded value.

The end results as far as the UI is concerned: Relations tables now show the name and type of nodes as well as IDs. The relations tables now only show placeholder relations if they are placeholders for THAT DIRECTION of relation, and never show node IDs, names or types for placeholders. Comments, names, types and IDs are now marked as "changed" only if they are different from the value stored in the DB, and it's now possible to drag data to the table using either the node ID or the name.