Added relation ordering; fixed a couple of bugs
My main project today was adding the ability to store relation orders. For now, I haven't done anything with logging in - I'm just assuming that whatever node ID is listed in the "User Node ID" field is the current user.

When you click "Save and Sync", the app looks for a relation between you (that is, the node whose ID is in the "User Node ID" field) and the node you're editing, which has the label "Order", and which has a "direction" value matching the specific relation table you're editing. If it finds one, it updates its "order" value to match the order of nodes you have created. If it doesn't find a matching "Order" relation, it creates one and gives it the appropriate "order" value.

When you create or refresh a relations table, it searches for all "Link" relations with that node, which will be shown on the table, but it ALSO searches for an "Order" relation between you and that node, with the "direction" value matching that relations table's direction value. If it finds an "Order" relation, it extracts the array of IDs which makes up the ordering.

It goes through those IDs one by one, searching for a matching node in the results of the relations query. If it finds a matching node, it adds it to the table and deletes it from the query results. Once it's processed all the IDs from the ordering results (which includes if there never were any ordering results), it adds the nodes which remain in the query results to the table.

Then I worked on the issue that data being added to the database weren't being escaped properly. I've enabled escaping when nodes are added or modified in widgetNode.add() and widgetNode.save(), and when relations are added or modified in widgetRelations.addNode() and widgetRelations.modifyNode(). I think those are the only functions that write to the DB, but if I remember any others I'll call stringEscape from those too. One caveat is that stringEscape doesn't escape single quotes, but if we use double quotes to surround string data when we write it to the database, it won't need to escape them - they're not treated as special characters when they're inside double quotes.

While I was doing that I spotted another bug. When I edited a node and then reopened it, I saw the old data. I eventually traced that back to the fact that widgetNode itself wasn't running a DB query when it was first opened - it was just taking the information that widgetTableNodes or widgetTableQuery passed it, which dated back to when the widgetTableNodes or widgetTableQuery widget was created. I rewrote widgetNode to query the DB whenever it was opened, in order to display accurate data. In order to do that, I rewrote widgetTableNodes, widgetTableQuery and widgetNodeTest to pass it an id instead of a whole node. I reran the test scripts for widgetNode to make sure it still worked; they all returned successes. I also made widgetNode display the node ID after a new node was added.