Mainly working on regression testing; starting to work on logging in
I wrote a simple login class, called widgetLogin. Right now it has no security - you just search for a name and, assuming there's a unique node with that name, you get logged in as that node. If there are no such nodes or more than one, you get an error message. We can expand it later, though, and it does enable you to log in without just typing an ID, which should make regression testing easier (I REALLY don't want to try to log IDs, since there's no way to guarantee they'll be the same). It's treated as a widget, just like regressionTesting, and builds its own header in the same way - it just needs an element with id="loginDiv".

I started working on writing tests for widgetRelations. I designed an HTML page and a test class, and played with them enough to be reasonably confident they work. In the process, I enabled logging of ALL the data returned when a new widgetNode is opened, and when a widgetNode or widgetRelations is refreshed. I have not written any widgetRelations test scripts yet - I expect to do that tomorrow.

I also took the logic I had been using in dbTest, to remove IDs from arbitrary data returned from a Cypher query, and moved it into a new function in app, stripIDs. I rewrote every line I could find that logged data from a Cypher query and rewrote it to make a copy of the data, if it didn't already, and to call the stripIDs function. It should now be easier to log data without the IDs, which are arbitrary and won't match from one run to the next.