More work on regression testing
Finished writing regression tests, except for app, because I couldn't think of anything to test on it which wasn't already tested in other classes' tests. Made a couple of changes to regressionTesting. Synthetic drag and drop events don't have a dataTransfer object, which I was using to drag data into widgetRelations, so I made a mock one so that drag and drop events could be reproduced.

I lost some time on a false trail looking for the best way to log the result of opening a node. It's pretty complex - from one click, you get data from FIVE separate queries (the node itself, its start and end relations, and the ordering of its start and end relations). I considered passing data around until it was all ready to log, or storing data in class variables (this.obj, for instance), for a lot longer than I should have before realizing the best thing to do was decouple the click and the data. I can log a click when it happens, and then log the data later, when it's available.

I made a small change to regressionTesting.record() so that if you're stepping through a recording and you get to a point where the next instruction doesn't involve an action - it's just data resulting from an action you already replayed - that instruction is "processed" without you needing to click through.

Found a bug I introduced when I started saving ordering - I was SETTING a query to add an ordering relation only if a user was logged in, but RUNNING it no matter what, with the weird-seeming result that ONLY IF no user was logged in, and the highest row in the table which needed processing was new or had its node changed, a duplicate relation was added. I moved the run query command into the if with the set query command and the bug went away.

Various small changes to facilitate logging: Added idrs to a few things that didn't have them and needed them; made the login controls log interactions so they could be recorded; made it log when you drag data from a widgetTableNodes to a widgetRelations to create a relation; made sure that when a node was opened, its start relation was finished before the end one was started, so they would log in order; added a userName value to widgetLogin and set it when a user logs in so there will be some way to refer to that user other than their ID; made the Trash table in widgetTableQuery not log IDs.

Created a very simple new class, "checkEmpty", which helps to check whether the database is empty while running a script. The pages which require an empty database all have a "Check whether the database is empty" button. When I record a script on these pages, the FIRST thing I do is click that button. It logs its interactions, so when that script is replayed, the first thing that happens is that button is "clicked". It calls a function in the checkEmpty class which runs a query to see all the nodes in the DB and passes the results to another function, verifyEmpty. If the DB is in fact empty, it just logs the empty array it got for data, and the recording continues. If the DB contains data, it logs the data, but it also displays an error message, stops the recording, resets all recording variables and removes the files the user uploaded. This button and the class that runs it are included in the test files for widgetTableNodes, widgetTableQuery and app.

Added instructions to the top of each HTML test page describing restrictions on when and how it can be run (needs an empty DB, early tests depend on later ones, has to have all tests run without refreshing, etc.)