Implemented personalized trashing
Rewrote code so people can trash nodes only for their own view. I removed the "trash" field, which no longer really made sense, and added a div to the bottom of each widgetNode which is visible only when a user is logged in, and which contains a checkbox marked "Trash Node" and a text input marked "Reason". The text input is only visible if you check the checkbox. If you check the box and save, a trash relation will be added from you, to your personal trash node, to the node you've just trashed (creating your trash node if necessary). If you uncheck the box and save, the relation from your trash node to that node will be removed.

I removed the generic "Trash" option from the metadata dropdown, and replaced it with "All trashed nodes" and "My trashed nodes". "My trashed nodes" is visible only when a user is logged in. If you click "My trashed nodes", you get basically the same thing you used to get when clicking "Trash" - it shows you all the nodes you have trashed, their labels and the reason you trashed them. I did edit it to also show the name.

If you click "All trashed nodes", you see a table of all nodes that ANYONE has trashed and the number of times each node has been trashed. Clicking on a node ID in this table opens a new table showing all the people who have trashed that node and why.

WidgetTableNodes now shows ALL nodes if you're not logged in, or the nodes you, personally, have trashed if you are.

Should I also rewrite widgetRelations so that it doesn't show relations to or from nodes that you have trashed? Also, can you remind me why we decided to have a trash node for each person? I remember agreeing on that, but now I can't remember why we said we couldn't have a Trash relation straight from the user to the node they wanted to trash.