Implementing personalized relations using views
Created a new widget type, widgetView, which appears where widgetRelations used to and shows a list of all people with views of a particular node in a particular direction. Each view has a toggle button which allows you to see the nodes that user listed. If you're logged in and have a view, it is opened automatically when you open the node. If you're logged in and do not have a view, there is an "Add Me" button which creates one and opens it for you.

Updated widgetRelation to use views instead of direct links between nodes, and to create two different types of tables depending on which view you're seeing. If you're looking at someone else's view, you can't change it, but you can drag nodes from it onto your own name in the View widget, which adds those nodes to your view. If you're looking at your own view, you can add, delete, change and rearrange nodes as before. Relations appear on the far side of the view widget (so on the left of the "end" widget and on the right of the "start" widget).

Also, added four things to widgetLogin that I think will really come in handy. viewLoggedIn and viewLoggedOut are arrays of DOM elements that are only visible when the user is logged in or logged out; widgetLogin updates their visibility whenever you log in or out. doOnLogin and doOnLogout are arrays of objects, where each object describes a function to be run: an object, object method, and array of parameters. The functions stored in that array are run when the user logs in or out. We can create as many widgets as we want, of as many different kinds as we want, and we should be able to update them all whenever the user logs in or out just by adding function descriptions to these arrays.