More work on logging in
Logging in now requires a username and password. People can be users or admins. Only admins can create new users or admins. When the page first loads, if no admin exists, it creates a temporary admin with username and password both "admin". If an admin does exist, the temporary admin is deleted.