Improved dragdrop functionality; enabled login on Enter
I rewrote some of the dragdrop functions. You should now be able to drag to your view (that is, the dragDrop table showing all the relations in your view) from the same view, another view or a table of nodes. Dragging from the same view rearranges the rows. Dragging from a table of nodes to the input row adds a new row; dragging from a table of nodes to any other row changes the node of that row. Dragging from another view adds a new row no matter what, and places it just before the row you dragged to. Also, I used to use what part of the row you dragged to to help to distinguish between rearranging rows and adding data, so I restricted which cells you could drag data to. That was stupid and I've fixed it. The new code records WHERE the data came from and uses that to decide what to do, and you can now drag data to any cell in the dragDrop table.

Finally, I got frustrated with having to click the "Log In" button every time I tested this stuff, so I added some code to widgetLogin so if you are in the password textbox, and you hit Enter, the program will log you in. I also added code to verify that the user has typed in both the username and password boxes - if they haven't, the program displays an alert telling them to enter their name and password, and doesn't bother contacting the DB and trying to log in.