Enabled rearranging children
You can now rearrange children by dragging them up or down. The rule I'm using for positioning is that if you drag a child even slightly higher than a sibling - that is, one pixel higher than if they perfectly overlapped - then it goes above that sibling. Since rearranging children only moves them vertically, I made them only move vertically when dragged, and if you mouse out - say, by dragging the mouse too far to the side - the child "snaps back" to where it was, the same way a scroll bar does if you move too far over. I was able to make an element's children, and the lines connecting them, move with it pretty easily. For now, I have NOT made the line connecting the element being dragged to its parent move, and I have NOT made its siblings move out of the way, because those would be more complicated to do and you said this doesn't need to be too polished just yet. I tested the code and found and fixed several bugs, and it seems to be working now, but if you find any errors, let me know.

I also created a small icon picture and saved it as favicon.ico to see if it would stop me from getting the error message saying favicon.ico wasn't found. It's hard to say for sure whether that helped, because I wasn't getting that message every time I loaded the page anyway, but I haven't gotten it since creating the file, so it may be helping. Anyway, I don't think it can hurt.