Still adding comments; found and fixed an inconsistency in widgetRelations
The behavior when you dropped data onto a dragDrop table vs. onto the
name of your view was not consistent - you could drag to the table from
your own view of something else, but you could only drag to the name of
your view from someone else’s view. I changed it so that you can drag
from your own view either way. I also found that dragging to the table
was not preserving comments, so I changed the code so that it does.