Still reviewing code; fixed bug in widgetView
When a new node was dragged into a table of relations using
widgetView’s drop function, its delete button was made to mark that row
for deletion by assigning it a class. That method of deletion was meant
for saved data, which can’t be truly deleted until the save button is
clicked and the DB is updated. New data doesn’t exist int he DB yet and
should simply be removed from the table - in fact, trying to delete it
from the DB (where it isn’t) causes an error. I changed the delete
button’s action to remove the new row from the table instead of marking
it for deletion.