More commenting, changes to graphic
Reviewed and commented a few more classes. Removed the idGet() method
from app as it was no longer needed. Updated how links work in the
graphic widget. You can now add a child to an element with no children
showing by dragging to the area in front of it (if it had hidden
children, the new child will appear at the end of the list). You can
also drag a new child above or below existing children to add it in
that position. Either of those methods will work both with existing
rectangles in the graphic and with data being dragged in from the
table. You can also drag existing rectangles directly onto other
rectangles to create a relationship, as before, but this won’t work
with data from the table, because we’re going to add a different
functionality when you drag from a table onto a rectangle.