Working on nodeless boxes
You can now double click the graphic to add a new box with a label
which is NOT connected to a node. For now, you can double-click
anywhere - eventually I will restrict that to only blank areas of the
graphic. For now, you can give a new box a name when you create it, but
you can’t change the name later, and you can’t link the box to a node
later. Eventually you will be able to do both of those.