Mainly working on graphic/mindmap
Added a few comments to dragDrop. Changed the name of “graphic” to
“mind map”. Fixed the bug where scrolling interfered with the placement
of the edit textbox. Introduced selected labels - whatever was most
recently clicked on is selected - will be useful later. Selected labels
currently have a green border but I expect the formatting to change
later. Enabled dragging a node from widgetTableNodes to a label to
connect them. Added a blue rectangle to the left of labels which are
linked to nodes. Mousing over the blue rectangle reveals a popup
containing details about that node, which update if a new node is
connected. Made both calendar and mindmap open the graphic immediately,
with a “Show Details” button to display the form beside the graphic;
created new “Details” class, heavily based on widgetNode, to handle
this. Made the work area only visible when logged in, and set it to
remove all widgets when the user logs out (so the next user won’t be
able to see the last one’s work).