Starting on keystrokes
Created a document-wide keydown listener. When it detects a keydown, it
checks whether the active widget a) exists and b) has a keydown
function. If so, it forwards the keydown event to the active widget (so
that each widget can process keystrokes in its own way). At the moment,
the only widget I plan to add keyboard shortcuts to is the mindmap, so
I gave it a keydown function. Right now all it does is alert that a key
was pressed - I’m about to start writing real functions for it to call.