Introduced keyboard shortcuts to mindmap
New shortcuts:

Enter creates a new younger sibling for the selected label, if the
selected label has a parent and the user isn’t editing a label (if they
are, Enter just means they’re done editing).

Tab creates a new child label for the selected label. If the selected
label has children already, the new one appears at the end of the list.
If the selected label’s children are hidden, they are displayed and the
new one is added as normal.

Delete deletes the selected label. For the time being, it also deletes
that label’s children. It does not select a new label - when you delete
a label, there is no selected label until you select a new one manually.

Escape deselects the selected label.

The arrow keys move through a tree. Left selects the selected label’s
parent, if any - if the selected label is a root it does nothing. Right
selects the selected label’s first child, if any - if the selected
label is a leaf, it does nothing. As with tab, if the selected label
has hidden children, they are shown first, and then the first one is
selected as normal.

Up and down cycle through the selected label’s siblings. If the
selected label’s is its parent’s last child, down will cycle around to
the first child, and if it’s its parent’s first child, up will cycle
around to the last.