Chrome compatibility
Changed onclick events of selects to onchange events because that seems
to work more reliably across browsers. Tested the most recent version
of Google Chrome, found no errors, and added it to the list of
supported browsers. Removed a few lines that I had commented out and
forgotten to delete.