Added "Show Node" and "Dissociate" buttons
Added two buttons to the node details popup in the mindmap. The one on
the left has an X in it, and if you click it, it removes the link to
that node. The popup, and the rectangle for showing it, disappear, and
the label’s name can be edited again. The button on the right has a +
in it, and if you click it, it opens a widgetNode form with all the
details about that node.

Also realized that clicking and dragging in the popup was moving the
label it was associated with. I don’t think that works well, and we
want to make the text in the popup selectable eventually, so I rewrote
the code so that only dragging the label itself moves it.