Implemented detaching labels
Uncle David suggested detaching child labels by just pulling them far
enough away from their parents, so I implemented that. “Far enough
away” at the moment is 50 px, but that is easy to change if it seems
too sensitive or not sensitive enough. The current parent gains or
loses its red highlighting based on whether the node being dragged is
close enough to “snap back” to it when released.