Bugfixes
Switched from using clientWidth to get the length of a text string, to
using getComputedTextLength(), which works in both Chrome and Safari.
Also replaced a couple of places where I built strings unnecessarily
with simple numbers, so that doing things like turning them negative
wouldn’t have unexpected side effects (- -20 is 20, but -“-20” is
apparently “—20”, which is bad if you want to parse it as a number).
Mindmaps appear to work in Chrome now.