Created data browser tool
There’s a new option in the metadata dropdown - data browser. When you
first open it, it’s a blank rectangle. Drag a node to it, and you get
two cells - the first shows details of that node, and the second shows
a list of links to it (like a widgetView, but with no filtering -
whatever’s in the DB is what you get). Click on a link, and you get two
more cells with details about the node on the other end and a list of
all ITS links, which you can click on, etc. The node(s) which are open
are highlighted in yellow. Clicking another node in a list which has
one open already replaces the open node; clicking the open one closes
it; dragging a node to the first cell starts the whole process over
using that new node. It’s less complicated than it sounds - try it out.