Better formatting on mindmap nodes
Updated button text to say “toggle” instead of “show”, so people will
be more likely to realize clicking the buttons again will make the item
they toggle disappear. Added white rectangles behind the explanation
popups so they are easier to read. Ensured that when an explanation
appears, it is in front of everything else so it can be read.