Bugfixes; preparation for DB move
Added lots of comments about DB functions in this and the last update.
Updated widgetSVG so that the delete key doesn’t delete nodes while
editing a textbox, so you can correct your work. Also fixed a bug in
which deleting a node while mousing over it caused its onmouseout event
to fire, and one in which deleting a node interfered with selecting
another node.