Implemented multiple selection
You can now select multiple nodes. If you just click an unselected
node, it is selected and all other nodes become unselected. If you hold
down Shift and click a node, its selected property toggles - it becomes
selected if it wasn’t already, and becomes deselected if it was
selected. If you hold down Shift and drag the mouse, a transparent
selection box appears; when you let go the box disappears and all
labels which it overlapped are selected. You can move multiple selected
nodes as a group, and delete them as a group. Other keyboard shortcuts
are disabled when multiple nodes are selected.