Drag files to mindmaps
You can now drag a file into a mindmap and create a node based on it.
The name is the name of the file, and the only detail is the type of
file. It is not yet possible to open the file - clicking the “+” button
just creates a popup saying we’re working on it.