Calendar updates
Removed the month and year from the day headers in the week view of the
calendar - now they only show the day of the week and day of the month.
Made the header text there a little bigger since there’s less of it to
show. Highlighted the selected view in yellow.